# meal-planner-system-frontend/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "operationId": "getById_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "operationId": "update_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCategoryRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "operationId": "delete_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "operationId": "getAllRoots",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CategoryResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "operationId": "create_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCategoryRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "refresh",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshTokenRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/recipes/slug/{slug}": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "operationId": "getBySlug",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RecipeResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/ingredients/slug/{slug}": {
            "get": {
                "tags": [
                    "Ingredients"
                ],
                "operationId": "getBySlug_1",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/IngredientResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/dietary/conditions/type/{type}": {
            "get": {
                "tags": [
                    "Dietary"
                ],
                "operationId": "getByType",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "CONTRAINDICATION",
                                "DIET"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DietaryConditionResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/categories/slug/{slug}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "operationId": "getBySlug_2",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "IngredientResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "normalizedName": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "unit": {
                        "type": "string",
                        "enum": [
                            "KG",
                            "G",
                            "L",
                            "ML",
                            "PCS",
                            "BUNCH"
                        ]
                    },
                    "stock": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "available": {
                        "type": "boolean"
                    },
                    "categoryId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "categoryName": {
                        "type": "string"
                    },
                    "aliases": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "UpdateCategoryRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "parentId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "CategoryResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "parentId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "parentName": {
                        "type": "string"
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CategoryResponse"
                        }
                    }
                }
            },
...
            "CreateCategoryRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 1
                    },
                    "parentId": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "required": [
                    "name"
                ]
            },
...
            "AuthResponse": {
                "type": "object",
                "properties": {
                    "accessToken": {
                        "type": "string"
                    },
                    "refreshToken": {
                        "type": "string"
                    },
                    "tokenType": {
                        "type": "string"
                    },
                    "expiresIn": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "RefreshTokenRequest": {
                "type": "object",
                "properties": {
                    "refreshToken": {
                        "type": "string"
                    }
                }
            },
...
            "RecipeIngredientResponse": {
                "type": "object",
                "properties": {
                    "ingredientId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "rawName": {
                        "type": "string"
                    },
                    "rawAmount": {
                        "type": "string"
                    },
                    "availableInShop": {
                        "type": "boolean"
                    },
                    "price": {
                        "type": "number"
                    },
                    "unit": {
                        "type": "string",
                        "enum": [
                            "KG",
                            "G",
                            "L",
                            "ML",
                            "PCS",
                            "BUNCH"
                        ]
                    },
                    "imageUrl": {
                        "type": "string"
                    }
                }
            },
            "RecipeResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "mealType": {
                        "type": "string"
                    },
                    "mealTypeDetailed": {
                        "type": "string"
                    },
                    "cookTime": {
                        "type": "string"
                    },
                    "cookComplexity": {
                        "type": "string"
                    },
                    "cookBudget": {
                        "type": "string"
                    },
                    "servings": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "servingSize": {
                        "type": "string"
                    },
                    "steps": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "ingredients": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RecipeIngredientResponse"
                        }
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "DietaryConditionResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "CONTRAINDICATION",
                            "DIET"
                        ]
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}