# meal-planner-system/src/main/java/com/feex/mealplannersystem/dto/dietary/DietaryConditionResponse.java
package com.feex.mealplannersystem.dto.dietary;

import com.feex.mealplannersystem.common.DietaryConditionType;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class DietaryConditionResponse {
    private String id;
    private String name;
    private String description;
    private DietaryConditionType type;
}